\name{getCacheFilename}
\alias{getCacheFilename}
\title{Returns the complete filepath to the cache file.}
\usage{
  getCacheFilename(query, dir = getwd(), ...)
}
\arguments{
  \item{query}{the query name.}

  \item{dir}{the directory to save the cache file to.}

  \item{...}{query parameters.}
}
\value{
  full filepath to the cached file.
}
\description{
  Returns the complete filepath to the cache file.
}

