\name{getParameters}
\alias{getParameters}
\title{Returns the parameters that must be set for the given query.}
\usage{
  getParameters(query)
}
\arguments{
  \item{query}{the query name.}
}
\value{
  list of parameter names.
}
\description{
  Returns the parameters that must be set for the given
  query.
}

