\name{getSQL}
\alias{getSQL}
\title{Returns the query as a string with the parameters set.}
\usage{
  getSQL(query = NULL, ...)
}
\arguments{
  \item{query}{the query name.}

  \item{...}{SQL parameters.}
}
\value{
  the SQL string with parameters replaced.
}
\description{
  Returns the query as a string with the parameters set.
}

