\name{parse.element}
\alias{parse.element}
\title{Parse a raw string containing key and expressions.}
\usage{
  parse.element(element, srcref)
}
\arguments{
  \item{element}{the string containing key and expressions}

  \item{srcref}{source reference.}
}
\value{
  A list containing the parsed constituents
}
\description{
  Copied from roxygen2:
  https://github.com/yihui/roxygen2/blob/master/R/parse-preref.R
}
\author{
  yihui
}

