\name{recodeColumns}
\alias{recodeColumns}
\title{Recodes factors with more than \code{maxLevels} to characters.}
\usage{
  recodeColumns(df, maxLevels = 20)
}
\arguments{
  \item{df}{the data frame to recode.}

  \item{maxLevels}{the maximum number of levels a factor
  can have before being converted to a character.}
}
\description{
  Recodes factors with more than \code{maxLevels} to
  characters.
}

