% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sqlexec.JDBCConnection}
\alias{sqlexec.JDBCConnection}
\title{Executes queries for RJDBC}
\usage{
\method{sqlexec}{JDBCConnection}(connection, sql, ...)
}
\arguments{
\item{connection}{the database connection.}

\item{sql}{the query to execute.}

\item{...}{other parameters passed to the appropriate \code{sqlexec} function.}
}
\description{
Executes queries for RJDBC
}

