% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sqlFile}
\alias{sqlFile}
\title{Returns the full path to the query or NULL if not found.}
\usage{
sqlFile(query)
}
\arguments{
\item{query}{the query to find.}
}
\value{
path to the query file.
}
\description{
Returns the full path to the query or NULL if not found.
}

