% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sqlPaths}
\alias{sqlPaths}
\title{Search paths for SQL repositories.}
\usage{
sqlPaths(path, replace = TRUE)
}
\arguments{
\item{path}{new path to add. This can a character vector of length greater than 1.}

\item{replace}{if FALSE, the path(s) will be added to already existing list.}
}
\description{
Search paths for SQL repositories.
}

