% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{sqlutils-package}
\alias{sqlutils}
\alias{sqlutils-package}
\title{Utilities for working with SQL files.}
\description{
Utilities for managing a library of SLQ files.
}
\author{
Jason Bryer \email{jason@bryer.org}
}
\keyword{database}
\keyword{package}
\keyword{sql}

