% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getParameters}
\alias{getParameters}
\title{Returns the parameters that must be set for the given query.}
\usage{
getParameters(query)
}
\arguments{
\item{query}{the query name.}
}
\value{
list of parameter names.
}
\description{
Returns the parameters that must be set for the given query.
}

