\name{getSQLRaw}
\alias{getSQLRaw}
\title{Returns the SQL from the file without the parameters replaced.}
\usage{
  getSQLRaw(query)
}
\arguments{
  \item{query}{the query name.}
}
\value{
  the unedited SQL statement.
}
\description{
  Returns the SQL from the file without the parameters
  replaced.
}

