\name{is.null.string}
\alias{is.null.string}
\title{Does the string contain no matter, but very well [:space:]?}
\usage{
  is.null.string(string)
}
\arguments{
  \item{string}{the string to check}
}
\value{
  TRUE if the string contains words, otherwise FALSE
}
\description{
  Does the string contain no matter, but very well
  [:space:]?
}

