% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sqlexec.RODBC}
\alias{sqlexec.RODBC}
\title{Executes queries for RODBC package.}
\usage{
\method{sqlexec}{RODBC}(connection, sql, ...)
}
\arguments{
\item{connection}{the database connection.}

\item{sql}{the query to execute.}

\item{...}{other parameters passed to the appropriate \code{sqlexec} function.}
}
\description{
Executes queries for RODBC package.
}

